using doubles, not ints...

+++ random.h
@@ -1,2 +1,2 @@
 // Yehuda Gale, 2016
-extern unsigned int random_maker(unsigned int random_number);
+extern double random_maker(double random_number);
